REFACTOR: rename blog list items
